# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Howick Cloud Server  - OpenAPI 3.0",
    "description": "This is Server for Howick Cloud Services for its customers using Howick machines. This server offer a long range of API's to communicate with machines and its web portals. You can find out more about Howick business at https://howickltd.com .\n\nHere is the documentation of API's used to interact with Server. A valid credentials are required to access server. ",
    "termsOfService": "https://www.howickltd.com/trading-terms",
    "contact": {
      "email": "support@howickltd.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.11"
  },
...
  "servers": [
    {
      "url": "http://localhost:5001/api/1.0.0/"
    }
  ],
...
  "paths": {
    "/assets": {
      "get": {
        "tags": [
          "asset"
        ],
        "summary": "get list of assets",
        "description": "Returns a list of assets",
        "operationId": "getAssets",
        "parameters": [
          
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Asset not found"
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "assets_auth": [
              "write:assets",
              "read:assets"
            ]
          }
        ]
      },
      "put": {
        "tags": [
          "asset"
        ],
        "summary": "Update an existing asset",
        "description": "Update an existing asset by Id",
        "operationId": "updateAsset",
        "requestBody": {
          "description": "Update an existent asset in the database",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Asset"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Asset"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Asset"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Asset not found"
          },
          "405": {
            "description": "Validation exception"
          }
        },
        "security": [
          {
            "assets_auth": [
              "write:assets",
              "read:assets"
            ]
          }
        ]
      },
      "post": {
        "tags": [
          "asset"
        ],
        "summary": "Add a new asset in DB",
        "description": "Add a new asset",
        "operationId": "addAsset",
        "requestBody": {
          "description": "Create a new asset",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Asset"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Asset"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Asset"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "assets_auth": [
              "write:assets",
              "read:assets"
            ]
          }
        ]
      }
    },
    "/assets/{assetId}": {
      "get": {
        "tags": [
          "asset"
        ],
        "summary": "Find asset by ID",
        "description": "Returns a single asset",
        "operationId": "getAssetById",
        "parameters": [
          {
            "name": "assetId",
            "in": "path",
            "description": "ID of asset to return",
            "required": true,
            "schema": {
              "type": "string",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Asset"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Asset not found"
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "assets_auth": [
              "write:assets",
              "read:assets"
            ]
          }
        ]
      },
      "patch": {
        "tags": [
          "asset"
        ],
        "summary": "Updates a asset in the DB with form data",
        "description": "",
        "operationId": "patchAsset",
        "parameters": [
          {
            "name": "assetId",
            "in": "path",
            "description": "ID of asset that needs to be updated",
            "required": true,
            "schema": {
              "type": "string",
              "format": "int64"
            }
          },
          {
            "name": "name",
            "in": "query",
            "description": "Name of asset that needs to be updated",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Status of asset that needs to be updated",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Status of asset that needs to be updated",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "assets_auth": [
              "write:assets",
              "read:assets"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "asset"
        ],
        "summary": "Deletes a asset",
        "description": "delete a asset",
        "operationId": "deleteAsset",
        "parameters": [
          {
            "name": "api_key",
            "in": "header",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assetId",
            "in": "path",
            "description": "Asset id to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid asset value"
          }
        },
        "security": [
          {
            "assets_auth": [
              "write:assets",
              "read:assets"
            ]
          }
        ]
      }
    },
...
    "/assetmodels/{modelId}": {
      "get": {
        "tags": [
          "model"
        ],
        "summary": "Find model by ID",
        "description": "Returns a single model",
        "operationId": "getAssetModelById",
        "parameters": [
          {
            "name": "modelId",
            "in": "path",
            "description": "ID of model to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetModel"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/AssetModel"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Asset Model not found"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "post": {
        "tags": [
          "model"
        ],
        "summary": "Updates a asset Model in the DB with form data",
        "description": "",
        "operationId": "patchAsset",
        "parameters": [
          {
            "name": "modelId",
            "in": "path",
            "description": "ID of asset Model that needs to be updated",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "name",
            "in": "query",
            "description": "Name of asset Model that needs to be updated",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Status of asset Model that needs to be updated",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        },
        "security": [
          {
            "assets_auth": [
              "write:assets",
              "read:assets"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "model"
        ],
        "summary": "Delete a model",
        "description": "delete a model",
        "operationId": "deleteAssetModel",
        "parameters": [
          {
            "name": "api_key",
            "in": "header",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assetId",
            "in": "path",
            "description": "Model id to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid model value"
          }
        },
        "security": [
          {
            "assets_auth": [
              "write:assets",
              "read:assets"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Dogs"
          }
        },
        "xml": {
          "name": "category"
        }
      },
...
      "Tag": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          }
        },
        "xml": {
          "name": "tag"
        }
      },
      "Asset": {
        "required": [
          "name",
          "photoUrls"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 10
          },
          "name": {
            "type": "string",
            "example": "doggie"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "photoUrls": {
            "type": "array",
            "xml": {
              "wrapped": true
            },
            "items": {
              "type": "string",
              "xml": {
                "name": "photoUrl"
              }
            }
          },
          "tags": {
            "type": "array",
            "xml": {
              "wrapped": true
            },
            "items": {
              "$ref": "#/components/schemas/Tag"
            }
          },
          "status": {
            "type": "string",
            "description": "asset status",
            "enum": [
              "available",
              "pending",
              "sold"
            ]
          }
        },
        "xml": {
          "name": "asset"
        }
      },
...
    },
...
    "securitySchemes": {
      "assets_auth": {
        "type": "oauth2",
        "flows": {
          "implicit": {
            "authorizationUrl": "http://localhost:3000/api/1.0.0/auth",
            "scopes": {
              "write:pets": "modify asset in your account",
              "read:pets": "read your assets"
            }
          }
        }
      },
      "api_key": {
        "type": "apiKey",
        "name": "api_key",
        "in": "header"
      }
    }
  }
}